+++ src/Dto/Widget.php
@@ -29,7 +29,8 @@ class Widget
         public readonly string $target,
         public readonly string $title,
         public readonly bool $expand,
-        public readonly IconType $iconType = IconType::FONT_AWESOME
+        public readonly IconType $iconType = IconType::FONT_AWESOME,
+        public readonly string $type = 'link'
     ) {
     }
 
@@ -47,6 +48,7 @@ class Widget
         $title = $data['title'] ?? '';
         $expand = $data['expand'] ?? false;
         $iconType = $data['icon_type'] ?? 'fa';
+        $type = $data['type'] ?? 'link';
 
         return new self(
             icon:     \is_string($icon) ? $icon : '',
@@ -55,7 +57,8 @@ class Widget
             target:   \is_string($target) ? $target : '',
             title:    \is_string($title) ? $title : '',
             expand:   \is_bool($expand) ? $expand : false,
-            iconType: IconType::tryFrom(\is_string($iconType) ? $iconType : 'fa') ?? IconType::FONT_AWESOME
+            iconType: IconType::tryFrom(\is_string($iconType) ? $iconType : 'fa') ?? IconType::FONT_AWESOME,
+            type:     \is_string($type) ? $type : 'link'
         );
     }
 }